+++ lib/just_color_picker.dart
@@ -1,4 +1,5 @@
 // Models
+export 'src/models/color_picker_type.dart';
 export 'src/models/color_state.dart';
 export 'src/models/color_format.dart';
 
@@ -7,6 +8,7 @@ export 'src/widgets/color_picker.dart';
 export 'src/widgets/hue_wheel.dart';
 export 'src/widgets/sv_panel.dart';
 export 'src/widgets/alpha_slider.dart';
+export 'src/widgets/hue_bar.dart';
 export 'src/widgets/hex_input.dart';
 export 'src/widgets/color_preview.dart';
 export 'src/widgets/color_info_panel.dart';